All: Fix Shadowing test failure when running all tests.
This patch ensures that each test is run with a clean environment. This required adding a new function to the interpreter to clear the environment. In addition to this function, the visibility of several classes had to be adjusted. This necessitated changes to the Expr and Stmt class code generation as well.